requirements.txt
asgiref==3.4.1
authy==2.2.6
certifi==2021.10.8
charset-normalizer==2.0.9
dj-database-url==0.5.0
Django==3.2.9
django-environ==0.8.1
django-heroku==0.3.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
environ==1.0
gunicorn==20.1.0
idna==3.3
psycopg2==2.9.2
PyJWT==2.3.0
python-dotenv==0.19.2
pytz==2021.3
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
twilio==7.3.2
urllib3==1.26.7
whitenoise==5.3.0
